requirements.txt
pygame==1.9.3
pyopengl==3.1.0
